arraytree with lambda based constructors
interestingly, eval and step are the same.

This approach makes it impossible to construct a program without fully
initializing both sides of an app/the body of a lambda, and the code
structure is a little more like the heap-tree approach compared to
arraytree but with lots of extra vertical bars.