statically prevent make_lam closure from allowing escape of arg
make_lam now requires the closure outlives the expr it produces, which
means that arguments to the closure can't include references that only
are valid within the closure.  This prevents sneaky things like
capturing a lambda's arg and using it as a value in the lambda
defining it.

For the Rc version, this means making the closure 'static, and for the
precise norc version it means introducing a new lifetime in make_lam.